rename codeceptjs in lib folder
