+++ lib/cli.js
@@ -19,7 +19,7 @@ class Cli extends Base {
     if (opts.debug) level = 2;
     if (opts.verbose) level = 3;
     output.level(level);
-    output.print(`CodeceptJS v${require('./codecept').version()}`);
+    output.print(`Hermiona v${require('./hermiona').version()}`);
     output.print(`Using test root "${global.codecept_dir}"`);
 
     const showSteps = level >= 1;
