+++ lib/command/definitions.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const path = require('path');
 
 const { getConfig, getTestRoot } = require('./utils');
-const Codecept = require('../codecept');
+const Codecept = require('../hermiona');
 const container = require('../container');
 const output = require('../output');
 const actingHelpers = [...require('../plugin/standardActingHelpers'), 'REST'];
