+++ lib/command/dryRun.js
@@ -1,6 +1,6 @@
 const { getConfig, getTestRoot } = require('./utils');
 const Config = require('../config');
-const Codecept = require('../codecept');
+const Codecept = require('../hermiona');
 const output = require('../output');
 const event = require('../event');
 const store = require('../store');
