+++ lib/command/gherkin/snippets.js
@@ -5,7 +5,7 @@ const glob = require('glob');
 const fsPath = require('path');
 
 const { getConfig, getTestRoot } = require('../utils');
-const Codecept = require('../../codecept');
+const Codecept = require('../../hermiona');
 const output = require('../../output');
 const { matchStep } = require('../../interfaces/bdd');
