+++ lib/command/gherkin/steps.js
@@ -1,5 +1,5 @@
 const { getConfig, getTestRoot } = require('../utils');
-const Codecept = require('../../codecept');
+const Codecept = require('../../hermiona');
 const output = require('../../output');
 const { getSteps } = require('../../interfaces/bdd');
