+++ lib/command/info.js
@@ -1,7 +1,7 @@
 const envinfo = require('envinfo');
 
 const { getConfig, getTestRoot } = require('./utils');
-const Codecept = require('../codecept');
+const Codecept = require('../hermiona');
 const output = require('../output');
 
 module.exports = async function (path) {
