+++ lib/command/interactive.js
@@ -1,6 +1,6 @@
 const { getConfig, getTestRoot } = require('./utils');
 const recorder = require('../recorder');
-const Codecept = require('../codecept');
+const Codecept = require('../hermiona');
 const event = require('../event');
 const output = require('../output');
