+++ lib/command/list.js
@@ -1,5 +1,5 @@
 const { getConfig, getTestRoot } = require('./utils');
-const Codecept = require('../codecept');
+const Codecept = require('../hermiona');
 const container = require('../container');
 const { getParamsToString } = require('../parser');
 const { methodsOfObject } = require('../utils');
