+++ lib/command/run-workers.js
@@ -25,7 +25,7 @@ module.exports = async function (workerCount, options) {
 
   const numberOfWorkers = parseInt(workerCount, 10);
 
-  output.print(`CodeceptJS v${require('../codecept').version()}`);
+  output.print(`Hermiona v${require('../hermiona').version()}`);
   output.print(`Running tests in ${output.styles.bold(numberOfWorkers)} workers...`);
   output.print();
