+++ lib/command/run.js
@@ -2,7 +2,7 @@ const {
   getConfig, printError, getTestRoot, createOutputDir,
 } = require('./utils');
 const Config = require('../config');
-const Codecept = require('../codecept');
+const Codecept = require('../hermiona');
 
 module.exports = async function (test, options) {
   // registering options globally to use in config
