+++ lib/command/workers/runTests.js
@@ -19,7 +19,7 @@ let stdout = '';
 const stderr = '';
 
 // Requiring of Codecept need to be after tty.getWindowSize is available.
-const Codecept = require(process.env.CODECEPT_CLASS_PATH || '../../codecept');
+const Codecept = require(process.env.CODECEPT_CLASS_PATH || '../../hermiona');
 
 const {
   options, tests, testRoot, workerIndex,
