+++ lib/index.js
@@ -7,9 +7,9 @@
  */
 module.exports = {
   /** @type {typeof CodeceptJS.Codecept} */
-  codecept: require('./codecept'),
+  codecept: require('./hermiona'),
   /** @type {typeof CodeceptJS.Codecept} */
-  Codecept: require('./codecept'),
+  Codecept: require('./hermiona'),
   /** @type {typeof CodeceptJS.output} */
   output: require('./output'),
   /** @type {typeof CodeceptJS.Container} */
