+++ lib/rerun.js
@@ -1,7 +1,7 @@
 const fsPath = require('path');
 const container = require('./container');
 const event = require('./event');
-const BaseCodecept = require('./codecept');
+const BaseCodecept = require('./hermiona');
 const output = require('./output');
 
 class CodeceptRerunner extends BaseCodecept {
