+++ lib/workers.js
@@ -5,7 +5,7 @@ const mkdirp = require('mkdirp');
 const { Worker } = require('worker_threads');
 const { Suite, Test, reporters: { Base } } = require('mocha');
 const ms = require('ms');
-const Codecept = require('./codecept');
+const Codecept = require('./hermiona');
 const MochaFactory = require('./mochaFactory');
 const Container = require('./container');
 const { getTestRoot } = require('./command/utils');
